+++ src/helix/bits/get_bits_leaderboard.rs
@@ -127,7 +127,7 @@ impl RequestGet for GetBitsLeaderboardRequest {
         request: Option<Self>,
         uri: &http::Uri,
         response: &str,
-        _: http::StatusCode,
+        status: http::StatusCode,
     ) -> Result<helix::Response<Self, Self::Response>, helix::HelixRequestGetError>
     where
         Self: Sized,
@@ -140,7 +140,12 @@ impl RequestGet for GetBitsLeaderboardRequest {
             total: i64,
         }
         let response: InnerResponse = helix::parse_json(response).map_err(|e| {
-            helix::HelixRequestGetError::DeserializeError(response.to_string(), e, uri.clone())
+            helix::HelixRequestGetError::DeserializeError(
+                response.to_string(),
+                e,
+                uri.clone(),
+                status,
+            )
         })?;
         Ok(helix::Response {
             data: BitsLeaderboard {