+++ src/helix/points.rs
@@ -33,6 +33,7 @@ pub mod create_custom_rewards;
 pub mod delete_custom_reward;
 pub mod get_custom_reward;
 pub mod get_custom_reward_redemption;
+pub mod update_custom_reward;
 pub mod update_redemption_status;
 
 #[doc(inline)]
@@ -45,6 +46,8 @@ pub use delete_custom_reward::{DeleteCustomReward, DeleteCustomRewardRequest};
 pub use get_custom_reward::{CustomReward, GetCustomRewardRequest};
 #[doc(inline)]
 pub use get_custom_reward_redemption::{CustomRewardRedemption, GetCustomRewardRedemptionRequest};
+#[doc(inline)]
+pub use update_custom_reward::{UpdateCustomRewardBody, UpdateCustomRewardRequest};
 #[doc(inline)]
 pub use update_redemption_status::{
     UpdateRedemptionStatusBody, UpdateRedemptionStatusInformation, UpdateRedemptionStatusRequest,