+++ src/helix/users/get_users_follows.rs
@@ -111,7 +111,7 @@ impl RequestGet for GetUsersFollowsRequest {
         request: Option<Self>,
         uri: &http::Uri,
         response: &str,
-        _status: http::StatusCode,
+        status: http::StatusCode,
     ) -> Result<helix::Response<Self, Self::Response>, helix::HelixRequestGetError>
     where
         Self: Sized,
@@ -125,7 +125,12 @@ impl RequestGet for GetUsersFollowsRequest {
         }
 
         let response: InnerResponse = helix::parse_json(response).map_err(|e| {
-            helix::HelixRequestGetError::DeserializeError(response.to_string(), e, uri.clone())
+            helix::HelixRequestGetError::DeserializeError(
+                response.to_string(),
+                e,
+                uri.clone(),
+                status,
+            )
         })?;
         Ok(helix::Response {
             data: UsersFollows {